Компоненты наследуются от абстрактного класса
