+++ src/view/point-view.js
@@ -1,6 +1,7 @@
 import dayjs from 'dayjs';
-import {createElement, getFormattedDuration} from '../utils/utils';
+import {getFormattedDuration} from '../utils/utils';
 import {TimeFormat} from '../consts';
+import AbstractView from './abstract-view';
 
 
 const getOffersTemplate = (offers) => `<h4 class="visually-hidden">Offers:</h4>
@@ -78,27 +79,15 @@ const createPointView = (point) => {
   </li>`;
 };
 
-export default class PointView {
+export default class PointView extends AbstractView {
   #point = null;
-  #element = null;
 
   constructor(point) {
+    super();
     this.#point = point;
   }
 
   get template() {
     return createPointView(this.#point);
-  }
-
-  get element() {
-    if(!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
